Update Bazel, rules_scala, and ScalaTest
