+++ .bazelversion
@@ -1,1 +1,1 @@
-4.2.0
+6.3.2
