+++ dependencies.yaml
@@ -87,19 +87,30 @@ dependencies:
   org.scalacheck:
     scalacheck:
       lang: scala
-      version: "1.13.5"
+      version: "1.17.0"
 
   org.scalactic:
     scalactic:
       lang: scala
-      version: "3.0.5"
+      version: "3.2.17"
 
   org.scalatest:
     scalatest:
       exports:
         - "org.scalactic:scalactic"
       lang: scala
-      version: "3.0.5"
+      version: "3.2.17"
+    scalatest-funsuite:
+      lang: scala
+      version: "3.2.17"
+    scalatest-propspec:
+      lang: scala
+      version: "3.2.17"
+
+  org.scalatestplus:
+    scalacheck-1-17:
+      lang: scala
+      version: "3.2.17.0"
 
   org.slf4j:
     slf4j:
