+++ pom.xml
@@ -254,7 +254,7 @@
     <dependency>
       <groupId>org.scala-lang</groupId>
       <artifactId>scala-reflect</artifactId>
-      <version>2.12.8</version>
+      <version>2.12.17</version>
     </dependency>
     <dependency>
       <groupId>org.scala-lang.modules</groupId>
@@ -269,12 +269,12 @@
     <dependency>
       <groupId>org.scalacheck</groupId>
       <artifactId>scalacheck_2.12</artifactId>
-      <version>1.13.5</version>
+      <version>1.17.0</version>
     </dependency>
     <dependency>
       <groupId>org.scalactic</groupId>
       <artifactId>scalactic_2.12</artifactId>
-      <version>3.0.5</version>
+      <version>3.2.17</version>
     </dependency>
     <dependency>
       <groupId>org.scalameta</groupId>
@@ -293,8 +293,83 @@
     </dependency>
     <dependency>
       <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-compatible</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-core_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-diagrams_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-featurespec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-flatspec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-freespec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-funspec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-funsuite_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-matchers-core_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-mustmatchers_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-propspec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-refspec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-shouldmatchers_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
+      <artifactId>scalatest-wordspec_2.12</artifactId>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatest</groupId>
       <artifactId>scalatest_2.12</artifactId>
-      <version>3.0.5</version>
+      <version>3.2.17</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scalatestplus</groupId>
+      <artifactId>scalacheck-1-17_2.12</artifactId>
+      <version>3.2.17.0</version>
     </dependency>
     <dependency>
       <groupId>org.slf4j</groupId>
