+++ src/scala/com/github/johnynek/bazel_deps/BUILD.bazel
@@ -158,10 +158,10 @@ scala_library(
     deps = [
         "//3rdparty/jvm/org/slf4j:slf4j_api",
         "//3rdparty/jvm/org/typelevel:cats_core",
-        "//3rdparty/jvm/org/typelevel:cats_free",
         "//3rdparty/jvm/org/typelevel:cats_effect",
         "//3rdparty/jvm/org/typelevel:cats_effect_kernel",
         "//3rdparty/jvm/org/typelevel:cats_effect_std",
+        "//3rdparty/jvm/org/typelevel:cats_free",
     ],
 )
 
@@ -187,8 +187,8 @@ scala_library(
     visibility = ["//visibility:public"],
     deps = [
         ":depsmodel",
-        ":graph",
         ":fs",
+        ":graph",
         "//3rdparty/jvm/io/circe:circe_core",
         "//3rdparty/jvm/io/circe:circe_generic",
         "//3rdparty/jvm/org/slf4j:slf4j_api",
@@ -214,9 +214,9 @@ scala_library(
         ":createpom",
         ":decoders",
         ":depsmodel",
+        ":fs",
         ":gradle_resolver",
         ":graph",
-        ":fs",
         ":normalizer",
         ":resolver",
         ":writer",
@@ -229,10 +229,10 @@ scala_library(
         "//3rdparty/jvm/org/eclipse/aether:aether_api",
         "//3rdparty/jvm/org/slf4j:slf4j_api",
         "//3rdparty/jvm/org/typelevel:cats_core",
-        "//3rdparty/jvm/org/typelevel:cats_free",
         "//3rdparty/jvm/org/typelevel:cats_effect",
-        "//3rdparty/jvm/org/typelevel:cats_effect_std",
         "//3rdparty/jvm/org/typelevel:cats_effect_kernel",
+        "//3rdparty/jvm/org/typelevel:cats_effect_std",
+        "//3rdparty/jvm/org/typelevel:cats_free",
     ],
 )
 
@@ -251,8 +251,8 @@ scala_binary(
         ":makedeps",
         "//3rdparty/jvm/org/slf4j:slf4j_simple",
         "//3rdparty/jvm/org/typelevel:cats_effect",
-        "//3rdparty/jvm/org/typelevel:cats_effect_std",
         "//3rdparty/jvm/org/typelevel:cats_effect_kernel",
+        "//3rdparty/jvm/org/typelevel:cats_effect_std",
     ],
 )
 
@@ -289,8 +289,8 @@ scala_library(
     visibility = ["//visibility:public"],
     deps = [
         ":depsmodel",
-        ":graph",
         ":fs",
+        ":graph",
         "//3rdparty/jvm/org/scala_lang/modules:scala_xml",
     ],
 )
